Fixing compilation errors
