+++ src/meldog_hardware/pi3hat_hardware_interface/include/controllers/ControllerBridge.hpp
@@ -3,41 +3,12 @@
 
 #include "../3rd_libs/pi3hat/pi3hat.h"
 #include "wrappers/ControllerWrapper.hpp"
+#include "ControllerStructures.hpp"
 #include <memory>
 #include <algorithm>
 
 namespace controller_interface
 {
-
-/* Structure for basic actuator command */
-struct ControllerCommand
-{
-    double position_;   /* [radians] */
-    double velocity_;   /* [radians/s] */
-    double torque_;     /* [Nm] */
-};
-
-/* Structure for basic actuator state */
-struct ControllerState
-{
-    double position_;   /* [radians] */
-    double velocity_;   /* [radians/s] */
-    double torque_;     /* [Nm] */
-    double temperature_;   /* [Celcius] */
-    bool fault = false;
-};
-
-struct ControllerParameters
-{
-    double position_max_;
-    double position_min_;
-    double position_offset_;
-    double velocity_max_;
-    double torque_max_;
-    int direction_ = 1;
-    int id_;             /* Usage in your bridge (check moteus bridge)*/
-    int bus_;            /* Usage in your bridge (check moteus bridge)*/
-};
 
 class ControllerBridge
 {
